+++ Sibusten.Philomena.Client/IPhilomenaImage.cs
@@ -29,6 +29,34 @@ namespace Sibusten.Philomena.Client
         string? OriginalHash { get; }
         List<string> TagNames { get; }
         List<int> TagIds { get; }
+        int? Score { get; }
+        int? FileSize { get; }
+        string? SourceUrl { get; }
+        bool? IsSpoilered { get; }
+        int? TagCount { get; }
+        bool? ThumbnailsGenerated { get; }
+        DateTime? UpdatedAt { get; }
+        string? Uploader { get; }
+        int? UploaderId { get; }
+        int? Upvotes { get; }
+        string? ViewUrl { get; }
+        bool? Processed { get; }
+        string? MimeType { get; }
+        bool? IsAnimated { get; }
+        double? AspectRatio { get; }
+        int? CommentCount { get; }
+        DateTime? CreatedAt { get; }
+        string? DeletionReason { get; }
+        string? Description { get; }
+        int? Downvotes { get; }
+        int? Width { get; }
+        int? DuplicateOf { get; }
+        int? Faves { get; }
+        DateTime? FirstSeenAt { get; }
+        int? Height { get; }
+        bool? IsHiddenFromUsers { get; }
+        double? Duration { get; }
+        double? WilsonScore { get; }
 
         /// <summary>
         /// Downloads the image